# Lots of bug fixes and performance improvements