--- a/ProjectLighthouse.Servers.GameServer/Controllers/ClientConfigurationController.cs
+++ b/ProjectLighthouse.Servers.GameServer/Controllers/ClientConfigurationController.cs
@@ -4,7 +4,6 @@ using System.Xml.Serialization;
 using LBPUnion.ProjectLighthouse.Configuration;
 using LBPUnion.ProjectLighthouse.PlayerData;
 using LBPUnion.ProjectLighthouse.PlayerData.Profiles;
-using LBPUnion.ProjectLighthouse.Types;
 using Microsoft.AspNetCore.Mvc;
 
 namespace LBPUnion.ProjectLighthouse.Servers.GameServer.Controllers;