--- a/ProjectLighthouse.Servers.GameServer/Controllers/Matching/EnterLevelController.cs
+++ b/ProjectLighthouse.Servers.GameServer/Controllers/Matching/EnterLevelController.cs
@@ -22,18 +22,15 @@ public class EnterLevelController : ControllerBase
     [HttpPost("play/user/{slotId}")]
     public async Task<IActionResult> PlayLevel(int slotId)
     {
-        User? user = await this.database.UserFromGameRequest(this.Request);
-        if (user == null) return this.StatusCode(403, "");
+        GameToken? token = await this.database.GameTokenFromRequest(this.Request);
+        if (token == null) return this.StatusCode(403, "");
 
         Slot? slot = await this.database.Slots.FirstOrDefaultAsync(s => s.SlotId == slotId);
         if (slot == null) return this.StatusCode(403, "");
 
-        GameToken? token = await this.database.GameTokenFromRequest(this.Request);
-        if (token == null) return this.StatusCode(403, "");
-
         GameVersion gameVersion = token.GameVersion;
 
-        IQueryable<VisitedLevel> visited = this.database.VisitedLevels.Where(s => s.SlotId == slotId && s.UserId == user.UserId);
+        IQueryable<VisitedLevel> visited = this.database.VisitedLevels.Where(s => s.SlotId == slotId && s.UserId == token.UserId);
         VisitedLevel? v;
         if (!visited.Any())
         {
@@ -49,9 +46,11 @@ public class EnterLevelController : ControllerBase
                 default: return this.BadRequest();
             }
 
-            v = new VisitedLevel();
-            v.SlotId = slotId;
-            v.UserId = user.UserId;
+            v = new VisitedLevel
+            {
+                SlotId = slotId,
+                UserId = token.UserId,
+            };
             this.database.VisitedLevels.Add(v);
         }
         else
@@ -88,13 +87,13 @@ public class EnterLevelController : ControllerBase
     [HttpGet("enterLevel/{id:int}")]
     public async Task<IActionResult> EnterLevel(int id)
     {
-        User? user = await this.database.UserFromGameRequest(this.Request);
-        if (user == null) return this.StatusCode(403, "");
+        GameToken? token = await this.database.GameTokenFromRequest(this.Request);
+        if (token == null) return this.StatusCode(403, "");
 
         Slot? slot = await this.database.Slots.FirstOrDefaultAsync(s => s.SlotId == id);
         if (slot == null) return this.NotFound();
 
-        IQueryable<VisitedLevel> visited = this.database.VisitedLevels.Where(s => s.SlotId == id && s.UserId == user.UserId);
+        IQueryable<VisitedLevel> visited = this.database.VisitedLevels.Where(s => s.SlotId == id && s.UserId == token.UserId);
         VisitedLevel? v;
         if (!visited.Any())
         {
@@ -103,7 +102,7 @@ public class EnterLevelController : ControllerBase
             v = new VisitedLevel
             {
                 SlotId = id,
-                UserId = user.UserId,
+                UserId = token.UserId,
             };
             this.database.VisitedLevels.Add(v);
         }