--- a/ProjectLighthouse.Servers.GameServer/Controllers/Resources/PhotosController.cs
+++ b/ProjectLighthouse.Servers.GameServer/Controllers/Resources/PhotosController.cs
@@ -2,6 +2,7 @@
 using System.Xml.Serialization;
 using Discord;
 using LBPUnion.ProjectLighthouse.Configuration;
+using LBPUnion.ProjectLighthouse.Extensions;
 using LBPUnion.ProjectLighthouse.Helpers;
 using LBPUnion.ProjectLighthouse.Levels;
 using LBPUnion.ProjectLighthouse.Logging;
@@ -63,7 +64,7 @@ public class PhotosController : ControllerBase
                 case SlotType.User:
                 {
                     Slot? slot = await this.database.Slots.FirstOrDefaultAsync(s => s.Type == SlotType.User && s.SlotId == photoSlot.SlotId);
-                    if (slot != null) validLevel = slot.RootLevel == photoSlot.RootLevel;
+                    if (slot != null && !string.IsNullOrEmpty(slot.RootLevel)) validLevel = true;
                     break;
                 }
                 case SlotType.Pod:
@@ -89,8 +90,19 @@ public class PhotosController : ControllerBase
 
         if (photo.Timestamp > TimeHelper.Timestamp) photo.Timestamp = TimeHelper.Timestamp;
 
+        // Check for duplicate photo subjects
+        List<string> subjectUserIds = new(4);
         foreach (PhotoSubject subject in photo.Subjects)
         {
+            if (subjectUserIds.Contains(subject.Username) && !string.IsNullOrEmpty(subject.Username)) return this.BadRequest();
+
+            subjectUserIds.Add(subject.Username);
+        }
+
+        foreach (PhotoSubject subject in photo.Subjects)
+        {
+            if (string.IsNullOrEmpty(subject.Username)) continue;
+
             subject.User = await this.database.Users.FirstOrDefaultAsync(u => u.Username == subject.Username);
 
             if (subject.User == null) continue;
@@ -103,18 +115,7 @@ public class PhotosController : ControllerBase
 
         await this.database.SaveChangesAsync();
 
-        // Check for duplicate photo subjects
-        List<int> subjectUserIds = new(4);
-        foreach (PhotoSubject subject in photo.Subjects)
-        {
-            if (subjectUserIds.Contains(subject.UserId)) return this.BadRequest();
-
-            subjectUserIds.Add(subject.UserId);
-        }
-
-        photo.PhotoSubjectIds = photo.Subjects.Select(subject => subject.PhotoSubjectId.ToString()).ToArray();
-
-        //            photo.Slot = await this.database.Slots.FirstOrDefaultAsync(s => s.SlotId == photo.SlotId);
+        photo.PhotoSubjectIds = photo.Subjects.Where(s => s.UserId != 0).Select(subject => subject.PhotoSubjectId.ToString()).ToArray();
 
         Logger.Debug($"Adding PhotoSubjectCollection ({photo.PhotoSubjectCollection}) to photo", LogArea.Photos);
 
@@ -203,7 +204,12 @@ public class PhotosController : ControllerBase
         Photo? photo = await this.database.Photos.FirstOrDefaultAsync(p => p.PhotoId == id);
         if (photo == null) return this.NotFound();
         if (photo.CreatorId != user.UserId) return this.StatusCode(401, "");
+        foreach (string idStr in photo.PhotoSubjectIds)
+        {
+            if (!int.TryParse(idStr, out int subjectId)) throw new InvalidCastException(idStr + " is not a valid number.");
 
+            this.database.PhotoSubjects.RemoveWhere(p => p.PhotoSubjectId == subjectId);
+        }
         this.database.Photos.Remove(photo);
         await this.database.SaveChangesAsync();
         return this.Ok();