… ProjectLighthouse.Servers.GameServer/Controllers/Slots/ListController.cs
… ProjectLighthouse.Servers.GameServer/Controllers/Slots/ReviewController.cs
… ProjectLighthouse.Servers.GameServer/Controllers/Slots/ScoreController.cs
… ProjectLighthouse.Servers.GameServer/Controllers/Slots/SearchController.cs
… ProjectLighthouse.Servers.GameServer/Controllers/Slots/SlotsController.cs
… ProjectLighthouse.Servers.GameServer/Startup/GameServerStartup.cs
… ProjectLighthouse.Servers.Website/Controllers/Debug/RoomVisualizerController.cs
… ProjectLighthouse.Servers.Website/Controllers/SlotPageController.cs
… ProjectLighthouse.Servers.Website/Controllers/UserPageController.cs
… ProjectLighthouse.Servers.Website/Pages/Admin/ModPanelPage.cshtml
… ProjectLighthouse.Servers.Website/Pages/CasePage.cshtml.cs
… ProjectLighthouse.Servers.Website/Pages/Debug/FilterTestPage.cshtml.cs
… ProjectLighthouse.Servers.Website/Pages/LandingPage.cshtml.cs
… ProjectLighthouse.Servers.Website/Pages/Partials/PhotoPartial.cshtml
… ProjectLighthouse.Servers.Website/Pages/PirateSignupPage.cshtml.cs
… ProjectLighthouse.Servers.Website/Pages/SlotPage.cshtml
… ProjectLighthouse.Servers.Website/Pages/SlotPage.cshtml.cs
… ProjectLighthouse.Servers.Website/Pages/SlotsPage.cshtml.cs
… ProjectLighthouse.sln.DotSettings
… ProjectLighthouse/Administration/Maintenance/MaintenanceJobs/CleanupUnusedPhotoSubjects.cs
… ProjectLighthouse/Administration/Maintenance/RepeatingTasks/CleanupRoomsTask.cs
… ProjectLighthouse/Database.cs
… ProjectLighthouse/Levels/Categories/HeartedCategory.cs
… ProjectLighthouse/Levels/Categories/NewestLevelsCategory.cs
… ProjectLighthouse/Levels/Categories/QueueCategory.cs
… ProjectLighthouse/Levels/Categories/TeamPicksCategory.cs
… ProjectLighthouse/Levels/Slot.cs
… ProjectLighthouse/Levels/VisitedLevel.cs
… ProjectLighthouse/Match/Rooms/RoomHelper.cs
… ProjectLighthouse/Migrations/20220808053519_RemoveVitaPlayCount.cs