This PR contains a bunch of stuff so I'm just gonna list the main points that I remember below
---
- You can no longer see sub-levels on the website if you aren't supposed to
- Pictures taken in sub-levels now show the proper level info on the website
- Spaces in the level search query now work properly on the website
- You can now upload photos taken with guests and online users not registered on your instance
- Tweaked serialization to be more efficient and behave more like the original servers
- PhotosWithMe has now been sped up considerably which should decrease load times of users and slots because the game often loads the slot author
- Changed Photos with me to only include photos that you are in and are not the creator of
- Removed LBPVita Playcount because a level can't be lbp2 and lbpv so we can just use lbp2 if it's vita.
- Added deflate compression similar to the original servers for responses larger than 1kb which should save a little bit of bandwidth on large responses
- crossControl slots are now detected (needs more testing) and properly serialized
- Fixed a bug where photo subjects would be added to the database even though the photo upload was rejected
- Removed including user location in User from GameToken as it was never actually used and just caused unnecessary DB calls
- the `/deleteReview` endpoint now requires authentication and level ownership
- Fixed cases where requests would accept 0 or negative page sizes which causes an error
- Fixed users not being able to rate comments in some cases in-game
- Use GameToken/Webtoken in lieu of User object in methods where the user wasn't modified to improve performance
- Projected id column instead of fetching whole objects from DB in certain cases.
- Made the `slots/developer` endpoint return comment and photocount because the game uses data from this request instead of checking each slot and without these counts it assumes that each slot has 0 of both.

Now that the boring stuff is out of the way let's get to the performance increases.

The Numbers
--- 
I ran some rudimentary tests that involved loading the game on RPCS3 and looking at the initial requests the game makes. I also played a level, looked at the tweak menu of a level link, and checked some options in the community tab. I first tested mainline and my PR with almost no photos, then I uploaded around 500 photos that each included 4 unique PhotoSubjects (2000 subjects and 500 photos). Lastly, I ran another test with 1000 photos and 4000 photo subjects. Another note is one or both of the PR tests I ran were with a version that didn't have all the optimizations of the final PR so latency might be even lower if reran.

A low number of photos (around 10)
--- 
These numbers didn't differ much between mainline and the PR
**Mainline**
```
 302ms: GET /LITTLEBIGPLANETPS3_XML/user/Slendy?version=2
 171ms: GET /LITTLEBIGPLANETPS3_XML/user/Slendy?version=2
 703ms: GET /LITTLEBIGPLANETPS3_XML/slots/by?u=Slendy&pageStart=1&pageSize=40&gameFilterType=both
  50ms: GET /LITTLEBIGPLANETPS3_XML/photos/user/56?by=Slendy&pageStart=1&pageSize=10
1054ms: GET /LITTLEBIGPLANETPS3_XML/slots/lbp2cool?pageStart=1&pageSize=30&gameFilterType=both&players=1&move=true
  12ms: GET /LITTLEBIGPLANETPS3_XML/slots/lolcatftw/Slendy?pageStart=1&pageSize=30&gameFilterType=both
```

A medium number of photos (around 500)
---
At this point, you could discern the latency even as an end user with basic user requests taking multiple seconds
**Mainline**
```
3839ms: GET /LITTLEBIGPLANETPS3_XML/user/Slendy?version=2
1812ms: GET /LITTLEBIGPLANETPS3_XML/slots/by?u=Slendy&pageStart=1&pageSize=40&gameFilterType=both
 330ms: GET /LITTLEBIGPLANETPS3_XML/slots/by?u=Slendy&pageStart=1&pageSize=1000&gameFilterType=both
1217ms: GET /LITTLEBIGPLANETPS3_XML/slots/lbp2luckydip?pageStart=1&pageSize=30&gameFilterType=both&seed=13
5555ms: GET /LITTLEBIGPLANETPS3_XML/user/Slendy?version=2
```
**PR**
```
 192ms: GET /LITTLEBIGPLANETPS3_XML/user/Slendy?version=2
 121ms: GET /LITTLEBIGPLANETPS3_XML/user/Slendy?version=2
1144ms: GET /LITTLEBIGPLANETPS3_XML/slots/by?u=Slendy&pageStart=1&pageSize=40&gameFilterType=both
 331ms: GET /LITTLEBIGPLANETPS3_XML/slots/by?u=Slendy&pageStart=1&pageSize=1000&gameFilterType=both
 134ms: GET /LITTLEBIGPLANETPS3_XML/slots/lbp2cool?pageStart=1&pageSize=30&gameFilterType=both&players=1&move=true
```
I also tested user times on the official Beacon instance and the results were closest to this category with the longest response time out of 4 users being 5496ms and the shortest being 3030ms

A high number of photos (around 1000)
---
At this point, the server felt practically unusable and I didn't test with numbers higher than this because I had a feeling it was going to keep scaling up with the number of photos and subjects
**Mainline**
```
 9350ms: GET /LITTLEBIGPLANETPS3_XML/user/Slendy?version=2
 1295ms: GET /LITTLEBIGPLANETPS3_XML/slots/by?u=Slendy&pageStart=1&pageSize=40&gameFilterType=both
  286ms: GET /LITTLEBIGPLANETPS3_XML/slots/by?u=Slendy&pageStart=1&pageSize=1000&gameFilterType=both
10595ms: GET /LITTLEBIGPLANETPS3_XML/user/Slendy?version=2
12319ms: GET /LITTLEBIGPLANETPS3_XML/photos/with?user=Slendy&pageStart=11&pageSize=10
  103ms: GET /LITTLEBIGPLANETPS3_XML/slots/lbp2cool?pageStart=1&pageSize=30&gameFilterType=both&players=1&move=true
```
**PR**
```
247ms: GET /LITTLEBIGPLANETPS3_XML/user/Slendy?version=2
 74ms: GET /LITTLEBIGPLANETPS3_XML/user/Slendy?version=2
456ms: GET /LITTLEBIGPLANETPS3_XML/slots/by?u=Slendy&pageStart=1&pageSize=40&gameFilterType=both
119ms: GET /LITTLEBIGPLANETPS3_XML/slots/by?u=Slendy&pageStart=1&pageSize=1000&gameFilterType=both
229ms: GET /LITTLEBIGPLANETPS3_XML/photos/with?user=Bruh&pageStart=1&pageSize=10
136ms: GET /LITTLEBIGPLANETPS3_XML/slots/lbp2luckydip?pageStart=1&pageSize=30&gameFilterType=both&seed=14
```

Closes #408 and closes #401