@TorutheRedFox 2022-08-09
shouldn't review deletion also allow deletion by the review's author?